ultimos

+++ proyecto/src/main/java/pe/edu/uni/proyecto/dto/RutaDto.java
@@ -24,5 +24,5 @@ public class RutaDto {
     private String destino;
 
     @JsonProperty("distancia")
-    private double distancia;
+    private long distancia;
 }
